misc/HtmlTestSuite/results/test-0800.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml">
<head>
<meta charset="utf-8"/>
<title>CSS Text Module Test: parsing tab-size with invalid values</title>
<link href="https://www.w3.org/TR/css-text-3/#propdef-tab-size" rel="help"/>
<meta content="tab-size supports only the grammar '&lt;number&gt; | &lt;length&gt;'." name="assert"/>
<meta content="Negative tab-size values are not allowed." name="assert"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="/css/support/parsing-testcommon.js"/>
</head>
<body>
<script>
test_invalid_value("tab-size", "-10px");
test_invalid_value("tab-size", "-20");
test_invalid_value("tab-size", "30%");
test_invalid_value("tab-size", "calc(40% + 50px)");
</script>
</body>
</html>
